requirements.txt
RouterOS-api
laiarturs-ros-api
rocketry
pydantic==1.10.9